backend/pyproject.toml
[project]
name = "hr-onboarding-agent"
version = "0.1.0"
description = "Employee Onboarding Agentic AI System"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.2.0",
    "azure-functions>=1.17.0",
    "azure-cosmos>=4.7.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.23.0",
    "pyright>=1.1.0",
    "ruff>=0.5.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
